fixes to general basis and also added tweeks to tools subpackage.
