+++ quspin/tools/measurements.py
@@ -7,27 +7,24 @@ import scipy.sparse.linalg as _sla
 import scipy.linalg as _la
 import numpy.linalg as _npla
 import scipy.sparse as _sp
-from scipy.special import binom, hyp2f1
 import numpy as _np
 from inspect import isgenerator as _isgenerator 
 
 # needed for isinstance only
 from ..operators import ishamiltonian as _ishamiltonian
 from ..operators import hamiltonian as _hamiltonian
-from ..basis import spin_basis_1d,photon_basis,isbasis
+from ..basis import isbasis as _isbasis
 from ..basis.photon import photon_Hspace_dim
+from .evolution import ED_state_vs_time
+from .misc import project_op,KL_div,mean_level_spacing
 
 import warnings
 
 
 __all__ =  ["ent_entropy", 
 			"diag_ensemble",
-			"ED_state_vs_time",
-			"obs_vs_time",
-			"project_op",
-			"KL_div",
-			"mean_level_spacing",
-			"evolve"]
+			"obs_vs_time"
+			]
 
 
 
@@ -712,7 +709,7 @@ def obs_vs_time(psi_t,times,Obs_dict,return_state=False,Sent_args={},enforce_pur
 		basis = Sent_args.get("basis")
 		if basis is None:
 			raise ValueError("Sent_args requires 'basis' for calculation")
-		if not isbasis(basis):
+		if not _isbasis(basis):
 			raise ValueError("'basis' object must be a proper basis object")
 
 		if ("chain_subsys" in Sent_args) or ("DM" in Sent_args) or ("svd_return_vec" in Sent_args):
@@ -785,572 +782,6 @@ def obs_vs_time(psi_t,times,Obs_dict,return_state=False,Sent_args={},enforce_pur
 			return_dict[i] = locals()[i]
 
 	return return_dict
-
-def ED_state_vs_time(psi,E,V,times,iterate=False):
-	"""Calculates the time evolution of initial state using a complete eigenbasis. 
-
-	The time evolution is carried out under the Hamiltonian :math:`H` with eigenenergies `E` and eigenstates `V`. 
-
-	Examples
-	--------
-
-	The following example shows how to time-evolve a state :math:`\\psi` using the entire eigensystem 
-	:math:`(E_1,V_1)` of a Hamiltonian :math:`H_1=\\sum_j hS^x_j + g S^z_j`.
-	
-	.. literalinclude:: ../../doc_examples/ED_state_vs_time-example.py
-		:linenos:
-		:language: python
-		:lines: 7-
-
-	Parameters
-	-----------
-	psi : numpy.ndarray
-		Initial state.
-	V : numpy.ndarray
-		Unitary matrix containing all eigenstates of the Hamiltonian :math:`H` in its columns. 
-	E : numpy.ndarray
-		Eigenvalues of the Hamiltonian :math:`H`, listed in the order which corresponds to the columns of `V`. 
-	times : numpy.ndarray
-		Vector of time to evaluate the time evolved state at. 
-	iterate : bool, optional
-		If set to `True`, the function returns the generator of the time evolved state. 
-
-	Returns
-	--------
-	obj
-		Either of the following:
-			* `numpy.ndarray` with the time evolved states as rows. 
-			* `generator` which generates time-dependent states one by one.
-
-	"""
-	psi = _np.squeeze(_np.asarray(psi))
-
-
-	if V.ndim != 2 or V.shape[0] != V.shape[1]:
-		raise ValueError("'V' must be a square matrix")
-
-	if V.shape[0] != len(E):
-		raise TypeError("Number of eigenstates in 'V' must equal number of eigenvalues in 'E'!")
-	if psi.shape[0] != len(E):
-		raise TypeError("Variables 'psi' and 'E' must have the same dimension!")
-
-	if psi.ndim == 2:
-		if psi.shape[0] != psi.shape[1]:
-			raise ValueError("mixed states must be square!")
-
-	if psi.ndim > 2:
-		raise ValueError("psi must be 1 or 2 dimension array.")
-
-	if _np.isscalar(times):
-		TypeError("Variable 'times' must be a array or iter like object!")
-
-	times = -1j*_np.asarray(times)
-	
-
-	# define generator of time-evolved state in basis V2
-	def pure_t_iter(V,psi,times):
-		# a_n: probability amplitudes
-		# times: time vector
-		a_n = V.T.conj().dot(psi)
-		for t in times:
-			yield V.dot( _np.exp(E*t)*a_n )
-
-	def mixed_t_iter(V,psi,times):
-		# a_n: probability amplitudes
-		# times: time vector
-		rho_d = V.T.conj().dot(psi.dot(V))
-		for t in times:
-			exp_t = _np.exp(t*E)
-			yield _np.einsum("ij,j,jk,k,lk->il",V,exp_t,rho_d,exp_t.conj(),V.conj())
-
-
-	if psi.ndim == 1:
-		if iterate:
-			return pure_t_iter(V,psi,times)
-		else:
-			c_n = V.T.conj().dot(psi)
-
-			Ntime = len(times)
-			Ns = len(E)
-
-			# generate [[-1j*times[0], ..., -1j*times[0]], ..., [-1j*times[-1], ..., -1j*times[01]]
-			psi_t = _np.broadcast_to(times,(Ns,Ntime)).T 
-			# [[-1j*E[0]*times[0], ..., -1j*E[-1]*times[0]], ..., [-1j*E[0]*times[-1], ..., -1j*E[-1]*times[-1]]
-			psi_t = psi_t*E
-			# [[exp(-1j*E[0]*times[0]), ..., exp(-1j*E[-1]*times[0])], ..., [exp(-1j*E[0]*times[-1]), ..., exp(-1j*E[01]*times[01])]
-			_np.exp(psi_t,psi_t)
-
-			# [[c_n[0]exp(-1j*E[0]*times[0]), ..., c_n[-1]*exp(-1j*E[-1]*times[0])], ..., [c_n[0]*exp(-1j*E[0]*times[-1]), ...,c_n[o]*exp(-1j*E[01]*times[01])]
-			psi_t *= c_n 
-
-			# for each vector trasform back to original basis
-			psi_t = V.dot(psi_t.T) 
-
-			return psi_t # [ psi(times[0]), ...,psi(times[-1]) ]
-	else:
-		if iterate:
-			return mixed_t_iter(V,psi,times)
-		else:
-			Ntime = len(times)
-			Ns = len(E)
-
-			rho_d = V.T.conj().dot(psi.dot(V))
-
-			# generate [[-1j*times[0], ..., -1j*times[0]], ..., [-1j*times[-1], ..., -1j*times[01]]
-			exp_t = _np.broadcast_to(times,(Ns,Ntime)).T
-			# [[-1j*E[0]*times[0], ..., -1j*E[-1]*times[0]], ..., [-1j*E[0]*times[-1], ..., -1j*E[-1]*times[-1]]
-			exp_t = exp_t*E
-			# [[exp(-1j*E[0]*times[0]), ..., exp(-1j*E[-1]*times[0])], ..., [exp(-1j*E[0]*times[-1]), ..., exp(-1j*E[01]*times[01])]
-			_np.exp(exp_t,exp_t)
-			
-			return _np.einsum("ij,tj,jk,tk,lk->ilt",V,exp_t,rho_d,exp_t.conj(),V.conj())
-
-def project_op(Obs,proj,dtype=_np.complex128):
-	"""Projects observable onto symmetry-reduced subspace.
-
-	This function takes an observable `Obs` and a reduced basis or a projector `proj`, and projects `Obs`
-	onto that reduced basis.
-
-	Examples
-	--------
-
-	The following example shows how to project an operator :math:`H_1=\\sum_j hS^x_j + g S^z_j` from the
-	symmetry-reduced basis to the full basis.
-	
-	.. literalinclude:: ../../doc_examples/project_op-example.py
-		:linenos:
-		:language: python
-		:lines: 7-
-
-	Parameters
-	-----------
-	Obs : :obj:
-		Operator to be projected, either a `numpy.ndarray` or a `hamiltonian` object.
-	proj : :obj:
-		Either one of the following:
-
-		* `basis` object with the basis of the Hilbert space after the projection.
-		* numpy.ndarray: a matrix which contains the projector.
-
-		Projectors can be calculated conveniently using the function method `basis.get_proj()`.
-	dtype : type, optional
-		Data type of output. Default is `numpy.complex128`.
-
-	Returns
-	-------- 
-	dict
-		Dictionary with keys
-
-		* "Proj_Obs": projected observable `Obs`.
-
-	"""
-
-	variables = ["Proj_Obs"]
-
-	if isbasis(proj):
-		proj = proj.get_proj(dtype)
-	elif (proj.__class__ not in [_np.ndarray,_np.matrix]) and (not _sp.issparse(proj)):
-		raise ValueError("Expecting either matrix/array or basis object for proj argument.")
-
-	if _ishamiltonian(Obs):
-
-		if Obs.Ns != proj.shape[0]:
-			if Obs.Ns != proj.shape[1]:
-				raise ValueError("Dimension mismatch Obs:{0} proj{1}".format(Obs.get_shape,proj.shape))
-			else:
-				# projecting from a smaller to larger H-space
-				proj_down=False
-		else:
-			# projecting from larger to smaller H-space
-			proj_down=True
-
-		if proj_down:
-			Proj_Obs = Obs.project_to(proj)		
-		else:
-			Proj_Obs = Obs.project_to(proj.T.conj())
-
-	else:
-
-		if Obs.ndim != 2:
-			raise ValueError("Expecting Obs to be a 2 dimensional array.")
-
-		if Obs.shape[0] != Obs.shape[1]:
-			raise ValueError("Expecting Obs to be a square array.")
-
-		if Obs.shape[1] != proj.shape[0]:
-			if Obs.shape[0] != proj.shape[1]:
-				raise ValueError("Dimension mismatch Obs:{0} proj{1}".format(Obs.shape,proj.shape))
-			else:
-				proj_down=False
-		else:
-			proj_down=True
-
-		if proj_down:
-			Proj_Obs = proj.T.conj().dot(Obs.dot(proj))
-		else:
-			Proj_Obs = proj.dot(Obs.dot(proj.T.conj()))
-
-	# define dictionary with outputs
-	return_dict = {}
-	for i in range(len(variables)):
-		return_dict[variables[i]] = locals()[variables[i]]
-
-	return return_dict
-
-def KL_div(p1,p2):
-	"""Calculates Kullback-Leibler divergence of two discrete probability distributions.
-
-	.. math::
-		\\mathrm{KL}(p_1||p_2) = \\sum_n p_1(n)\\log\\frac{p_1(n)}{p_2(n)}
-
-	Parameters
-	----------- 
-	p1 : numpy.ndarray
-		Dscrete probability distribution.
-	p2 : numpy.ndarray
-		Discrete probability distribution.
-
-	Returns
-	--------
-	numpy.ndarray
-		Kullback-Leibler divergence of `p1` and `p2`.
-
-	"""
-	p1 = _np.asarray(p1)
-	p2 = _np.asarray(p2)
-
-
-	if len(p1) != len(p2):
-		raise TypeError("Expecting the probability distributions 'p1' and 'p2' to have same size!")
-	if p1.ndim != 1 or p2.ndim != 1:
-		raise TypeError("Expecting the probability distributions 'p1' and 'p2' to have linear dimension!")
-
-
-	if _np.any(p1<=0.0) or _np.any(p2<=0.0):
-		raise TypeError("Expecting all entries of the probability distributions 'p1' and 'p2' to be non-negative!")
-	
-	if abs(sum(p1)-1.0) > 1E-13:
-		raise ValueError("Expecting 'p1' to be normalised!")
-
-	if abs(sum(p2)-1.0) > 1E-13:
-		raise ValueError("Expecting 'p2' to be normalised!")
-
-	if _np.any(p1==0.0):
-
-		inds = _np.where(p1 == 0)
-
-		p1 = _np.delete(p1,inds)
-		p2 = _np.delete(p2,inds)
-
-	return _np.multiply( p1, _np.log( _np.divide(p1,p2) ) ).sum()
-
-def mean_level_spacing(E):
-	"""Calculates the mean-level spacing of an energy spectrum.
-
-	See mean level spacing, :math:`\\langle\\tilde r_\mathrm{W}\\rangle`, in 
-	`arXiv:1212.5611 <https://arxiv.org/pdf/1212.5611.pdf>`_ for more details.
-
-	For Wigner-Dyson statistics, we have :math:`\\langle\\tilde r_\mathrm{W}\\rangle\\approx 0.53`, while
-	for Poisson statistics: :math:`\\langle\\tilde r_\mathrm{W}\\rangle\\approx 0.38`.
-
-	Examples
-	--------
-
-	The following example shows how to calculate the mean level spacing :math:`\\langle\\tilde r_\mathrm{W}\\rangle` for the
-	spectrum of the ergodic Hamiltonian :math:`H_1=\\sum_jJ S^z_{j+1}S^z + hS^x_j + g S^z_j`.
-	
-	.. literalinclude:: ../../doc_examples/mean_level_spacing-example.py
-		:linenos:
-		:language: python
-		:lines: 7-
-
-	Parameters
-	-----------
-	E : numpy.ndarray
-		Ordered list of ascending, NONdegenerate energies.
-
-	Returns
-	-------- 
-	float
-		mean-level spacing.
-
-	"""
-
-	if not isinstance(E,_np.ndarray):
-		E = _np.asarray(E)
-
-	if _np.any(_np.sort(E)!=E):
-		raise TypeError("Expecting a sorted list of ascending, nondegenerate eigenenergies 'E'.")
-
-	# compute consecutive E-differences
-	sn = _np.diff(E)
-	# check for degeneracies
-	if len(_np.unique(E)) != len(E):
-		raise ValueError("Degeneracies found in spectrum 'E'!")
-	# calculate the ratios of consecutive spacings
-	aux = _np.zeros((len(E)-1,2),dtype=_np.float64)
-
-	aux[:,0] = sn
-	aux[:,1] = _np.roll(sn,-1)
-
-	return _np.mean(_np.divide( aux.min(1), aux.max(1) )[0:-1] )
-
-
-##### below are the routines for arbitary user-defimed time evolution.
-
-def evolve(v0,t0,times,f,solver_name="dop853",real=False,stack_state=False,verbose=False,imag_time=False,iterate=False,f_params=(),**solver_args):
-	"""Implements (imaginary) time evolution for a user-defined first-order ODE.
-
-	The function can be used to study nonlinear semiclassical dynamics. It can also serve as a pre-configured 
-	ODE solver in python, without any relation to other QuSpin objects.
-
-	Examples
-	--------
-
-	The following example shows how to use the`evolve()` function to solve the periodically-driven 
-	Gross-Pitaevskii equation (GPE) on a one-imensional lattice. The GPE has a linear part, comprising the
-	kinetic energy and the external potentials (e.g. a harmonic trap), and a nonlinear part which describes
-	the interactions.
-
-	Below, in a few steps we show how to use the functionality of the `evolve()` function to solve the GPE 
-	on a one-dimensional lattice for periodically-driven particles in a harmonic trap:
-
-	.. math::
-		i\\dot\\varphi_j(t) = -J\\left( e^{-iA\\sin\\Omega t}\\varphi_{j-1}(t) + e^{+iA\\sin\\Omega t}\\varphi_{j+1}(t) \\right) + \\kappa_\\mathrm{trap}\\varphi_j(t) + U|\\varphi_j(t)|^2\\varphi_j(t)
-
-	where :math:`j` labels the lattice sites, :math:`J` is the lattice hopping amplitude, :math:`\\kappa_\\mathrm{trap}` is
-	the strength of the harmonic trap, and :math:`U` -- the mean-field interaction.
-
-	Let us start by defining the single-particle Hamiltonian :math:`H(t)`.
-
-
-	.. literalinclude:: ../../doc_examples/evolve-example.py
-		:linenos:
-		:language: python
-		:lines: 7-47
-
-	Next, we define the GPE 
-
-	.. math::
-		-i\\dot\\varphi(t) = H(t)\\varphi(t) + U |\\varphi(t)|^2 \\varphi(t)
-
-	and solve it using `evolve()`:
-
-	.. literalinclude:: ../../doc_examples/evolve-example.py
-		:linenos:
-		:language: python
-		:lines: 49-61
-		:lineno-start: 43
-
-	Since the GPE is a complex-valued equation, the above code requires the use of a complex-valued ODE solver 
-	[which is done by `evolve()` under the hood, so long as no solver is explicitly specified].
-
-	An alternative way 
-	to solve the GPE using a real-valued solver might be useful to speed-up the computation. This can be achieved 
-	by decomposing the condensate wave function into a real and imaginary part, and proceeds as follows:
-	
-	The goal is to solve: 
-	
-	.. math::
-		-i\\dot\\varphi(t) = H(t)\\varphi(t) + U |\\varphi(t)|^2 \\varphi(t)
-	
-	for the complex-valued :math:`\\varphi(t)` by re-writing it as a real-valued vector `phi=[u,v]` where 
-	:math:`\\varphi(t) = u(t) + iv(t)`. The real and imaginary parts, :math:`u(t)` and :math:`v(t)`, have the same dimension as 
-	:math:`\phi(t)`.
-	
-	In the most general form, the single-particle Hamiltonian can be decomposed as 
-	:math:`H(t)= H_{stat} + f(t)H_{dyn}`, with a complex-valued driving function :math:`f(t)`. Then, the GPE can be cast in 
-	the following real-valued form:
-	
-	.. math::
-		\\dot u(t) = +\\left[H_{stat} + U\\left(|u(t)|^2 + |v(t)|^2\\right) \\right]v(t) + Re[f(t)]H_{dyn}v(t) + Im[f(t)]H_{dyn}u(t)
-	.. math::	
-		\\dot v(t) = -\\left[H_{stat} + U\\left(|u(t)|^2 + |v(t)|^2\\right) \\right]u(t) - Re[f(t)]H_{dyn}u(t) + Im[f(t)]H_{dyn}v(t)
-
-
-	.. literalinclude:: ../../doc_examples/evolve-example.py
-		:linenos:
-		:language: python
-		:lines: 63-
-		:lineno-start: 59
-
-	The flag `stack_state=True` is required for `evolve()` to handle the complex-valued initial condition properly,
-	as well as to put together the output solution as a complex-valued vector in the end. Since the real-valued ODE solver 
-	allows to parse ODE parameters, we can include them in the user-defined ODE function and use the 
-	flag `f_params`. Notice the elegant way python allows one to circumvent the usage of this variable in the 
-	complex-valued example above. 
-
-	Parameters
-	-----------
-	v0 : numpy.ndarray
-		Initial state.
-	t0 : float
-		Initial time.
-	times : numpy.ndarray
-		Vector of times to compute the time-evolved state at.
-	f : :obj:`function`
-		User-defined function to solve first-order ODE (see Examples):
-
-		.. math::
-			v'(t) = f(v(t),t)\\qquad v(t_0)=v_0
-	f_params : tuple, optional
-		A tuple to pass all parameters of the function `f` to ODE solver. Default is `f_params=()`.
-	iterate : bool, optional
-		If set to `True`, creates a generator object for the time-evolved the state. Default is `False`.
-	solver_name : str, optional
-		Scipy solver integrator name. Default is `dop853`. 
-
-		See `scipy integrator (solver) <https://docs.scipy.org/doc/scipy-0.14.0/reference/generated/scipy.integrate.ode.html>`_ for other options.
-	solver_args : dict, optional
-		Dictionary with additional `scipy integrator (solver) <https://docs.scipy.org/doc/scipy-0.14.0/reference/generated/scipy.integrate.ode.html>`_ arguments.	
-	real : bool, optional 
-		Flag to determine if `f` is real or complex-valued. Default is `False`.
-	imag_time : bool, optional
-		Must be set to `True` when `f` defines imaginary-time evolution, in order to normalise the state 
-		at each time in `times`. Default is `False`.
-	stack_state : bool, optional
-		If `f` is written to take care of real and imaginary parts separately (see Examples), this flag 
-		will return a single complex-valued solution instead of the real and imaginary parts separately.
-		Default is `False`.
-	verbose : bool, optional
-		If set to `True`, prints normalisation of state at teach time in `times`.
-
-	Returns
-	--------
-	obj
-		Can be either one of the following:
-			* numpy.ndarray containing evolved state against time.
-			* generator object for time-evolved state (requires `iterate = True`).
-		
-	"""
-
-	from scipy.integrate import complex_ode
-	from scipy.integrate import ode
-
-	if v0.ndim > 2:
-		raise ValueError("state mush have ndim < 3.")
-
-	shape0 = v0.shape
-
-	if v0.ndim == 2:
-		if v0.shape[0] != v0.shape[1]:
-			v0 = v0.ravel()
-	 
-	
-	if _np.iscomplexobj(times):
-		raise ValueError("times must be real number(s).")
-
-	n = _np.linalg.norm(v0) # needed for imaginary time to preserve the proper norm of the state. 
-
-	if stack_state:
-		v1 = v0
-		v0 = _np.zeros(v0.shape[0],dtype=v1.real.dtype)
-		v0[:v0.shape[0]] = v1.real
-		v0[v0.shape[0]:] = v1.imag
-
-		solver = ode(f) # y_f = f(t,y,*args)
-	elif real:
-		solver = ode(f) # y_f = f(t,y,*args)
-	else:
-		solver = complex_ode(f) # y_f = f(t,y,*args)
-
-	if solver_name in ["dop853","dopri5"]:
-		if solver_args.get("nsteps") is None:
-			solver_args["nsteps"] = _np.iinfo(_np.int32).max
-		if solver_args.get("rtol") is None:
-			solver_args["rtol"] = 1E-9
-		if solver_args.get("atol") is None:
-			solver_args["atol"] = 1E-9
-
-
-	solver.set_integrator(solver_name,**solver_args)
-	solver.set_f_params(*f_params)
-	solver.set_initial_value(v0, t0)
-
-	if _np.isscalar(times):
-		return _evolve_scalar(solver,v0,t0,times,stack_state,imag_time,n,shape0)
-	else:
-		if iterate:
-			return _evolve_iter(solver,v0,t0,times,verbose,stack_state,imag_time,n,shape0)
-		else:
-			return _evolve_list(solver,v0,t0,times,verbose,stack_state,imag_time,n,shape0)
-
-
-def _evolve_scalar(solver,v0,t0,time,stack_state,imag_time,n,shape0):
-	from numpy.linalg import norm
-	Ns=shape0[0]
-
-	if time == t0:
-		if stack_state:
-			return (v0[:Ns] + 1j*v0[Ns:]).reshape(shape0)
-		else:
-			return _np.array(v0).reshape(shape0)
-
-	solver.integrate(time)
-	if solver.successful():
-		if imag_time: solver._y /= (norm(solver._y)/n)
-		if stack_state:
-			return (solver.y[:Ns] + 1j*solver.y[Ns:]).reshape(shape0)
-		else:
-			return _np.array(solver.y).reshape(shape0)
-	else:
-		raise RuntimeError("failed to evolve to time {0}, nsteps might be too small".format(time))	
-
-def _evolve_list(solver,v0,t0,times,verbose,stack_state,imag_time,n,shape0):
-	from numpy.linalg import norm
-
-	Ns=shape0[0]
-	v = _np.empty(shape0+(len(times),),dtype=_np.complex128)
-	
-	for i,t in enumerate(times):
-
-		if t == t0:
-			if verbose: print("evolved to time {0}, norm of state {1}".format(t,_np.linalg.norm(solver.y)))
-			if stack_state:
-				v[...,i] = (v0[:Ns] + 1j*v0[Ns:]).reshape(shape0)
-			else:
-				v[...,i] = _np.array(v0).reshape(shape0)
-			continue
-
-		solver.integrate(t)
-		if solver.successful():
-			if verbose: print("evolved to time {0}, norm of state {1}".format(t,_np.linalg.norm(solver.y)))
-			if imag_time: solver._y /= (norm(solver._y)/n)
-			if stack_state:
-				v[...,i] = (solver.y[:Ns] + 1j*solver.y[Ns:]).reshape(shape0)
-			else:
-				v[...,i] = solver.y.reshape(shape0)
-		else:
-			raise RuntimeError("failed to evolve to time {0}, nsteps might be too small".format(t))
-			
-
-	return v
-
-def _evolve_iter(solver,v0,t0,times,verbose,stack_state,imag_time,n,shape0):
-	from numpy.linalg import norm
-	Ns=shape0[0]
-
-
-	for i,t in enumerate(times):
-		if t == t0:
-			if verbose: print("evolved to time {0}, norm of state {1}".format(t,_np.linalg.norm(solver.y)))
-			if stack_state:
-				yield (v0[:Ns] + 1j*v0[Ns:]).reshape(shape0)
-			else:
-				yield _np.array(v0).reshape(shape0)
-			continue
-			
-
-		solver.integrate(t)
-		if solver.successful():
-			if verbose: print("evolved to time {0}, norm of state {1}".format(t,_np.linalg.norm(solver.y)))
-			if imag_time: solver._y /= (norm(solver._y)/n)
-			if stack_state:
-				yield (solver.y[:Ns] + 1j*solver.y[Ns:]).reshape(shape0)
-			else:
-				yield solver.y.reshape(shape0)
-		else:
-			raise RuntimeError("failed to evolve to time {0}, nsteps might be too small".format(t))
 
 ##### private functions
